project restructured: - services from docker compose file moved to docker compose file in parent repository - FastAPI routers now in dedicated folder - web page removed completely
