+++ Dockerfile
@@ -2,10 +2,10 @@ FROM python:3.11-slim-buster
 
 WORKDIR /app
 
-COPY .requirements_symink ./requirements.txt
+COPY requirements.txt requirements.txt
 
 RUN pip install --no-cache-dir -r requirements.txt
 
-COPY src ./python/src
+COPY python/src ./python/src
 
 CMD ["python", "python/src/FastAPI/main.py"]
