--- python/.requirements_symink
@@ -1,138 +1,0 @@
-# This file was autogenerated by uv via the following command:
-#    uv pip compile - -o requirements.txt
-annotated-types==0.7.0
-    # via pydantic
-anyio==4.3.0
-    # via
-    #   httpx
-    #   starlette
-    #   watchfiles
-attrs==23.2.0
-    # via
-    #   cattrs
-    #   requests-cache
-cattrs==23.2.3
-    # via requests-cache
-certifi==2024.2.2
-    # via
-    #   httpcore
-    #   httpx
-    #   requests
-charset-normalizer==3.3.2
-    # via requests
-click==8.1.7
-    # via
-    #   typer
-    #   uvicorn
-dnspython==2.6.1
-    # via email-validator
-email-validator==2.1.1
-    # via fastapi
-fastapi==0.111.0
-fastapi-cli==0.0.4
-    # via fastapi
-flatbuffers==24.3.25
-    # via openmeteo-sdk
-geographiclib==2.0
-    # via geopy
-geopy==2.4.1
-h11==0.14.0
-    # via
-    #   httpcore
-    #   uvicorn
-httpcore==1.0.5
-    # via httpx
-httptools==0.6.1
-    # via uvicorn
-httpx==0.27.0
-    # via fastapi
-idna==3.7
-    # via
-    #   anyio
-    #   email-validator
-    #   httpx
-    #   requests
-jinja2==3.1.4
-    # via fastapi
-markdown-it-py==3.0.0
-    # via rich
-markupsafe==2.1.5
-    # via jinja2
-mdurl==0.1.2
-    # via markdown-it-py
-multimethod==1.11.2
-numpy==1.26.4
-    # via pandas
-openmeteo-requests==1.2.0
-openmeteo-sdk==1.11.4
-    # via openmeteo-requests
-orjson==3.10.3
-    # via fastapi
-pandas==2.2.2
-platformdirs==4.2.2
-    # via requests-cache
-psycopg2-binary==2.9.9
-pydantic==2.7.1
-    # via fastapi
-pydantic-core==2.18.2
-    # via pydantic
-pygments==2.18.0
-    # via rich
-python-dateutil==2.9.0.post0
-    # via pandas
-python-dotenv==1.0.1
-    # via uvicorn
-python-multipart==0.0.9
-    # via fastapi
-pytz==2024.1
-    # via pandas
-pyyaml==6.0.1
-    # via uvicorn
-requests==2.31.0
-    # via
-    #   openmeteo-requests
-    #   requests-cache
-    #   retry-requests
-requests-cache==1.2.0
-retry-requests==2.0.0
-rich==13.7.1
-    # via typer
-shellingham==1.5.4
-    # via typer
-six==1.16.0
-    # via
-    #   python-dateutil
-    #   url-normalize
-sniffio==1.3.1
-    # via
-    #   anyio
-    #   httpx
-starlette==0.37.2
-    # via fastapi
-typer==0.12.3
-    # via fastapi-cli
-typing-extensions==4.11.0
-    # via
-    #   fastapi
-    #   pydantic
-    #   pydantic-core
-    #   typer
-tzdata==2024.1
-    # via pandas
-ujson==5.10.0
-    # via fastapi
-url-normalize==1.4.3
-    # via requests-cache
-urllib3==2.2.1
-    # via
-    #   requests
-    #   requests-cache
-    #   retry-requests
-uvicorn==0.29.0
-    # via fastapi
-uvloop==0.19.0
-    # via uvicorn
-watchfiles==0.21.0
-    # via uvicorn
-websockets==12.0
-    # via uvicorn
