+++ python/src/FastAPI/config.py
@@ -1,0 +1,4 @@
+import base64
+import os
+
+JWT_PUBLIC_KEY = os.environ['JWT_PUBLIC_KEY']
