+++ python/src/FastAPI/main.py
@@ -1,18 +1,13 @@
-import os
 from fastapi import FastAPI
 
 from smartphone_routers import router as smartphone_router
-from users import router as users_router
-from researches import router as researches_router
-from kits import router as kit_router
-from samples import router as samples_router
+from routers.users import router as users_router
+from routers.researches import router as researches_router
+from routers.kits import router as kit_router
+from routers.samples import router as samples_router
 
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi import FastAPI, Form, Request
-from fastapi.responses import HTMLResponse
-from fastapi.templating import Jinja2Templates
-from fastapi.staticfiles import StaticFiles
-
+from fastapi import FastAPI
 app = FastAPI()
 
 origins = ["*"]
@@ -30,36 +25,6 @@ app.include_router(researches_router)
 app.include_router(kit_router)
 app.include_router(samples_router)
 app.include_router(smartphone_router)
-
-
-
-# Mount the static files directory
-app.mount("/static", StaticFiles(directory="python/src/FastAPI/static"), name="static")
-
-
-@app.get("/", response_class=HTMLResponse)
-async def read_root():
-    html_path = "python/src/FastAPI/templates/index.html"
-    with open(html_path, 'r') as file:
-        html_content = file.read()
-    return HTMLResponse(content=html_content)
-
-
-@app.get("/admin/sign_in", response_class=HTMLResponse)
-async def admin_sign_in():
-    html_path = "python/src/FastAPI/templates/admin_sign_in.html"
-    with open(html_path, 'r') as file:
-        html_content = file.read()
-    return HTMLResponse(content=html_content)
-
-
-templates = Jinja2Templates(directory="python/src/FastAPI/templates")
-@app.post("/admin/sign_in", response_class=HTMLResponse)
-async def handle_admin_sign_in(request: Request, login: str = Form(...), password: str = Form(...)):
-    if login == "admin" and password == "password":
-        return templates.TemplateResponse("admin_page.html", {"request": request})
-    else:
-        return templates.TemplateResponse("admin_sign_in.html", {"request": request, "error": "Invalid credentials"})
 
 
 if __name__ == "__main__":
