--- python/src/FastAPI/templates/admin_page.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Biokeeper</title>
-
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
-    <link rel="stylesheet" href="/static/style.css">
-
-</head>
-<body>
-    <div class="container center-align">
-        <img src="/static/biokeeper_logo.png" alt="Biokeeper Logo" class="logo">
-        <h1>Biokeeper</h1>
-        <p>You are logged in!</p>
-
-        <div class="button-row">
-            <a href="/" class="btn-large waves-effect waves-light">Go back</a>
-        </div>
-    
-    </div>
-</body>
-</html>
