--- python/src/FastAPI/templates/admin_sign_in.html
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Biokeeper - Login</title>
-    
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
-    <link rel="stylesheet" href="/static/style.css">
-    
-</head>
-<body>
-    <div class="container center-align">
-        <img src="/static/biokeeper_logo.png" alt="Biokeeper Logo" class="logo">
-        <h1>Biokeeper Admin</h1>
-        <form action="/admin/sign_in" method="post" class="login-form">
-            <div class="input-field">
-                <input id="login" name="login" type="text" class="validate">
-                <label for="login">Login</label>
-            </div>
-            <div class="input-field">
-                <input id="password" name="password" type="password" class="validate">
-                <label for="password">Password</label>
-            </div>
-            <div class="button-row">
-                <a href="/" class="btn-large waves-effect waves-light">Go back</a>
-                <button type="submit" class="btn-large waves-effect waves-light">Sign In</button>
-            </div>
-        </form>
-    </div>
-
-    <!-- Import jQuery and Materialize JS -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
-    <script>
-        $(document).ready(function() {
-            M.updateTextFields(); // Initialize Materialize components
-
-            // Add 'active' class on focus
-            $('input').on('focus', function() {
-                $(this).siblings('label').addClass('active');
-            });
-
-            // Remove 'active' class on blur if input is empty
-            $('input').on('blur', function() {
-                if (!$(this).val()) {
-                    $(this).siblings('label').removeClass('active');
-                }
-            });
-
-            // Check input fields on page load
-            $('input').each(function() {
-                if ($(this).val()) {
-                    $(this).siblings('label').addClass('active');
-                }
-            });
-        });
-    </script>
-</body>
-</html>
